added docs for particle, and split lithium-ion tests by model
