+++ docs/source/models/submodels/particle.rst
@@ -1,0 +1,5 @@
+Particle Submodel
+==================
+
+.. autoclass:: pybamm.particle.Standard
+    :members:
